backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
supabase>=2.4.0
langchain>=0.1.20
langchain-community>=0.0.38
langchain-core>=0.1.52
langchain-huggingface>=0.0.3
transformers>=4.38.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.1